yt-whisper/pyproject.toml
[tool.poetry]
name = "yt-whisper"
version = "0.1.0"
description = ""
authors = ["David Cardozo <david.cardozo@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.29.0"
python-dotenv = "^1.0.0"
pytube = "^15.0.0"
openai = "^1.7.0"
pinecone-client = "^2.2.4"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
